polish: 코드 정리
