+++ src/main/java/wooteco/subway/controller/LineController.java
@@ -3,9 +3,9 @@ package wooteco.subway.controller;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import wooteco.subway.controller.request.LineAndSectionCreateRequest;
 import wooteco.subway.controller.response.LineCreateResponse;
 import wooteco.subway.controller.response.LineWithAllSectionsResponse;
-import wooteco.subway.controller.request.LineAndSectionCreateRequest;
 import wooteco.subway.service.LineService;
 import wooteco.subway.service.dto.LineDto;
