+++ src/main/java/wooteco/subway/controller/StationController.java
@@ -2,10 +2,9 @@ package wooteco.subway.controller;
 
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.WebDataBinder;
 import org.springframework.web.bind.annotation.*;
-import wooteco.subway.controller.response.StationResponse;
 import wooteco.subway.controller.request.StationRequest;
+import wooteco.subway.controller.response.StationResponse;
 import wooteco.subway.service.StationService;
 import wooteco.subway.service.dto.StationDto;
