+++ src/test/java/wooteco/subway/dao/station/StationDaoTest.java
@@ -16,7 +16,7 @@ import java.util.stream.Collectors;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static wooteco.subway.dao.fixture.CommonFixture.makeStation;
+import static wooteco.subway.dao.fixture.Fixture.makeStation;
 
 @JdbcTest
 public class StationDaoTest {
